# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rolling",
    "description": "롤링페이퍼 서비스",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/rollingpapers": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "sender": {
                  "example": "any"
                },
                "receiver": {
                  "example": "any"
                },
                "created_at": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "롤링페이퍼 작성이 완료되었습니다."
          },
          "400": {
            "description": "롤링페이퍼 작성을 실패했습니다."
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/rollingpapers/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/papers": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "paper_id": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "writer": {
                  "example": "any"
                },
                "font": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/papers/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/papers/{paper_id}/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/users/{email}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "403": {
            "description": "금지됨"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "권한 없음"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/admin": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/vip": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    }
  }
}